requirements.txt
streamlit>=1.28.0
openai-whisper>=20231117
torch>=2.0.0
torchaudio>=2.0.0
transformers>=4.21.0
soundfile>=0.12.1
numpy>=1.21.0
ffmpeg-python>=0.2.0
